pyproject.toml
[tool.poetry]
name = "edu"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.6"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
django-storages = "^1.14.2"
boto3 = "^1.28.73"
djangorestframework = "^3.14.0"
celery = "^5.3.5"
rabbitmq-server = "^0.0.1"
factory-boy = "^3.3.0"
faker = "^20.0.0"
redis = "^5.0.1"
model-bakery = "^1.17.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
model-bakery = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
